+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
  * Add PHP attributes to register listeners and guards
  * Deprecate `GuardEvent::getContext()` method that will be removed in 7.0
  * Revert: Mark `Symfony\Component\Workflow\Registry` as internal
+ * Add `WorkflowGuardListenerPass` (moved from `FrameworkBundle`)
 
 6.2
 ---